+++ tests_corpus/valid/primaruy-key-collation-sorting.sql
@@ -1,4 +1,4 @@
-CREATE TABLE person(
+CREATE TABLE personF(
     fullname NOT NULL,
     birthday DEFAULT NULL,
     PRIMARY KEY (fullname COLLATE BINARY ASC)